\name{emp}
\docType{data}
\alias{emp}
\title{Sample data for RHive examples}
\description{
This data set contains employee information such as their name, age, department and position.
}
\usage{data(emp)}
\format{emp is a data frame with 9 rows and 8 columns named 
  id, ename, position, dep, enterdate, sal, status, and age.}
\keyword{datasets}
