\name{rhive-fn}
\alias{hiveConnect}
\alias{hiveQuery}
\alias{hiveExport}
\alias{hiveExportAll}
\alias{hiveClose}
\alias{hiveListTables}
\alias{hiveDescTable}
\alias{hiveLoadTable}
\alias{hiveAssign}
\alias{hiveRm}
\title{Methods for the class }
\description{ this functions are wrapper for rhive functions to provide method style.}
\author{
\email{rhive@nexr.com}
}
\keyword{programming}
