% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/udaf.R
\name{write_udaf_scripts}
\alias{write_udaf_scripts}
\title{Writes User Defined Aggregation Function}
\usage{
write_udaf_scripts(f, cluster_by, input_table, input_cols, output_table,
  output_cols, base_name = "udaf", sep = "\\t")
}
\arguments{
\item{f}{function which accepts a grouped data frame and returns a
data frame}

\item{cluster_by}{character name of column to \code{CLUSTER BY}}

\item{input_table}{character name of table to be transformed, ie.
\code{SELECT input_cols FROM input_table}. Can also contain more SQL,
such as \code{input_table WHERE col1 < 10}.}

\item{input_cols}{character vector of input columns to f}

\item{output_table}{character name of table to \code{INSERT INTO
output_table}}

\item{output_cols}{character vector of columns that f will output}

\item{base_name}{character base name of script to write ie. foo.R and foo.sql}

\item{sep}{field separator string}
}
\value{
scripts character vector containing generated scripts
}
\description{
Generates R and SQL scripts to call as user defined aggregation
functions in Hive
}
\details{
Feedback:

Do I attempt to have consistency with similar funcs /
packages? Ie. DBI package uses statement, lapply uses FUN

Alternatively I could use caps to denote SQL things, ie. CLUSTER_BY
}
\examples{
write_udaf_scripts(
}
