\name{rhive-fn}
\alias{hiveConnect}
\alias{hiveQuery}
\alias{hiveExport}
\alias{hiveExportAll}
\alias{hiveClose}
\alias{hiveListDatabases}
\alias{hiveListTables}
\alias{hiveShowDatabases}
\alias{hiveUseDatabase}
\alias{hiveShowTables}
\alias{hiveDescTable}
\alias{hiveLoadTable}
\alias{hiveAssign}
\alias{hiveRm}
\title{Methods for the class }
\description{ These functions are wrappers to provide function style api 
  for rhive functions.}
\author{
\email{rhive@nexr.com}
}
\keyword{programming}
