\name{rhive}
\alias{RHive}
\alias{rhive}
\docType{package}
\title{rhive: R and Hive}
\description{
	RHive is an R extension facilitating distributed computing via HIVE query.
	It allows easy usage of HQL in R, and allows easy usage of R objects and R functions in Hive.
}
\author{
\email{rhive@nexr.com}
}
\details{
  The rhive package supplies functions to interact
  with Hive from within R. 
  There are functions for exporting and connecting
  as well as querying hive.
} 
\seealso{
rhive.init
rhive.connect
rhive.query
rhive.assign
rhive.rm
rhive.export
rhive.exportAll
rhive.script.export
rhive.script.unexport
rhive.close
rhive.list.tables
rhive.desc.table
rhive.load.table
rhive.napply
rhive.sapply
rhive.aggregate
rhive.mrapply
rhive.mapapply
rhive.reduceapply
rhive.hdfs.connect
rhive.hdfs.ls
rhive.hdfs.get
rhive.hdfs.put
rhive.hdfs.rm
rhive.hdfs.rename
rhive.hdfs.exists
rhive.hdfs.mkdirs
rhive.hdfs.close
hiveConnect
hiveQuery
hiveAssign
hiveRm
hiveExport
hiveExportAll
hiveClose
hiveListTables
hiveDescTable
hiveLoadTable
}
\keyword{programming}
\concept{RHive}
